Change initialization. User controller is added.
